# different strategy for tokenizing keywords