I did my idea from https://github.com/ziglang/zig/pull/5359#issuecomment-634171853

throughput: 279 MiB/s => 347 MiB/s