@andrewrk 2020-05-26
benchmark code:

```zig
const std = @import("std");

pub fn setup(gpa: *std.mem.Allocator, options: *Options) !void {}
var global_total_byte_count: u64 = undefined;
var global_total_token_count: u64 = undefined;

pub fn run(gpa: *std.mem.Allocator, context: void) !void {
    var walker = try std.fs.walkPath(gpa, "../lib/std");
    defer walker.deinit();

    var total_byte_count: u64 = 0;
    var total_token_count: u64 = 0;
    while (try walker.next()) |entry| switch (entry.kind) {
        .File => {
            if (!std.mem.endsWith(u8, entry.basename, ".zig")) continue;

            const source = try entry.dir.readFileAlloc(gpa, entry.basename, 30 * 1024 * 1024);
            defer gpa.free(source);

            total_byte_count += source.len;

            var tokenizer = std.zig.Tokenizer.init(source);
            while (true) {
                const token = tokenizer.next();
                total_token_count += 1;
                if (token.id == .Eof) break;
            }

        },
        else => continue,
    };

    global_total_byte_count = total_byte_count;
    global_total_token_count = total_token_count;
}

pub const Measurement = struct {
    median: u64,
    mean: u64,
    min: u64,
    max: u64,

    fn compute(all_samples: []Sample, comptime field: []const u8) Measurement {
        const S = struct {
            fn order(a: Sample, b: Sample) bool {
                return @field(a, field) < @field(b, field);
            }
        };
        // Remove the 2 outliers
        std.sort.sort(Sample, all_samples, S.order);
        const samples = all_samples[1 .. all_samples.len - 1];
        // Compute stats
        var total: u64 = 0;
        var min: u64 = std.math.maxInt(u64);
        var max: u64 = 0;
        for (samples) |s| {
            const v = @field(s, field);
            total += v;
            if (v < min) min = v;
            if (v > max) max = v;
        }
        return .{
            .median = @field(samples[samples.len / 2], field),
            .mean = total / samples.len,
            .min = min,
            .max = max,
        };
    }
};

pub const Results = union(enum) {
    fail: anyerror,
    ok: struct {
        samples_taken: usize,
        wall_time: Measurement,
        utime: Measurement,
        stime: Measurement,
        maxrss: usize,
    },
};

const Sample = struct {
    wall_time: u64,
    utime: u64,
    stime: u64,
};

fn timeval_to_ns(timeval: std.os.timeval) u64 {
    const ns_per_us = std.time.ns_per_s / std.time.us_per_s;
    return @bitCast(usize, timeval.tv_sec) * std.time.ns_per_s +
        @bitCast(usize, timeval.tv_usec) * ns_per_us;
}

var samples_buf: [1000000]Sample = undefined;
const max_nano_seconds = std.time.ns_per_s * 20;

pub fn bench(options: Options, comptime func: var, args: var) Results {
    var sample_index: usize = 0;
    const timer = std.time.Timer.start() catch @panic("need timer to work");
    const first_start = timer.read();
    while ((sample_index < 3 or
        (timer.read() - first_start) < max_nano_seconds) and
        sample_index < samples_buf.len)
    {
        const start_rusage = std.os.getrusage(options.rusage_who);
        const start = timer.read();
        @call(.{}, func, args) catch |err| {
            return .{ .fail = err };
        };
        const end = timer.read();
        const end_rusage = std.os.getrusage(options.rusage_who);
        samples_buf[sample_index] = .{
            .wall_time = end - start,
            .utime = timeval_to_ns(end_rusage.utime) - timeval_to_ns(start_rusage.utime),
            .stime = timeval_to_ns(end_rusage.stime) - timeval_to_ns(start_rusage.stime),
        };
        sample_index += 1;
    }
    const all_samples = samples_buf[0..sample_index];
    const wall_time = Measurement.compute(all_samples, "wall_time");
    const utime = Measurement.compute(all_samples, "utime");
    const stime = Measurement.compute(all_samples, "stime");

    const final_rusage = std.os.getrusage(options.rusage_who);

    std.debug.warn("total_token_count={}\ntotal_byte_count={}\n", .{global_total_token_count, global_total_byte_count});
    {
        const elapsed_s = @intToFloat(f64, wall_time.mean) / std.time.ns_per_s;
        const throughput = @floatToInt(u64, @intToFloat(f64, global_total_byte_count) / elapsed_s);
        std.debug.warn("throughput (mean): {Bi}/s\n", .{throughput});
    }
    {
        const elapsed_s = @intToFloat(f64, wall_time.median) / std.time.ns_per_s;
        const throughput = @floatToInt(u64, @intToFloat(f64, global_total_byte_count) / elapsed_s);
        std.debug.warn("throughput (median): {Bi}/s\n", .{throughput});
    }
    {
        const elapsed_s = @intToFloat(f64, wall_time.max) / std.time.ns_per_s;
        const throughput = @floatToInt(u64, @intToFloat(f64, global_total_byte_count) / elapsed_s);
        std.debug.warn("throughput (min): {Bi}/s\n", .{throughput});
    }
    {
        const elapsed_s = @intToFloat(f64, wall_time.min) / std.time.ns_per_s;
        const throughput = @floatToInt(u64, @intToFloat(f64, global_total_byte_count) / elapsed_s);
        std.debug.warn("throughput (max): {Bi}/s\n", .{throughput});
    }

    return .{
        .ok = .{
            .samples_taken = all_samples.len,
            .wall_time = wall_time,
            .utime = utime,
            .stime = stime,
            .maxrss = @bitCast(usize, final_rusage.maxrss),
        },
    };
}

pub const Options = struct {
    rusage_who: i32 = std.os.RUSAGE_SELF,
};

pub fn main() !void {
    const gpa = if (std.builtin.link_libc) std.heap.c_allocator else std.heap.page_allocator;
    var options: Options = .{};
    const context = try setup(gpa, &options);
    const results = bench(options, run, .{ gpa, context });
    try std.json.stringify(results, std.json.StringifyOptions{}, std.io.getStdOut().outStream());
}
```

run instructions:

```
./zig run benchmark.zig --release-fast -lc
```

master:

```
throughput (mean): 285.748743057251MiB/s
throughput (median): 285.61156463623047MiB/s
throughput (min): 276.09959506988525MiB/s
throughput (max): 292.609414100647MiB/s
{"samples_taken":342,"wall_time":{"median":58652599,"mean":58624442,"min":57249903,"max":60673253},"utime":{"median":56092000,"mean":55936076,"min":51945000,"max":58686000},"stime":{"median":2979000,"mean":2686317,"min":978000,"max":6031000},"maxrss":280412}
```

comptime hashmap:

```
throughput (mean): 334.53428840637207MiB/s
throughput (median): 333.30327892303467MiB/s
throughput (min): 326.11223888397217MiB/s
throughput (max): 341.7137613296509MiB/s
{"samples_taken":400,"wall_time":{"median":50271341,"mean":50086354,"min":49034030,"max":51379865},"utime":{"median":47398000,"mean":47257206,"min":43324000,"max":49551000},"stime":{"median":2985000,"mean":2827223,"min":971000,"max":6981000},"maxrss":299060}
```

this branch:

```
throughput (mean): 345.59295654296875MiB/s
throughput (median): 345.70947074890137MiB/s
throughput (min): 333.6345806121826MiB/s
throughput (max): 363.7453269958496MiB/s
{"samples_taken":413,"wall_time":{"median":48459932,"mean":48476270,"min":46057107,"max":50213792},"utime":{"median":45891000,"mean":45711233,"min":41717000,"max":48545000},"stime":{"median":2963000,"mean":2756588,"min":957000,"max":7042000},"maxrss":33812}
```