@andrewrk 2020-05-26
I think it might be useful to start collecting instruction counts rather than wall clock timings because the wall clock timings tend to be highly variable for me